Move relations and expressions into their own modules
