--- datafusion/substrait/src/logical_plan/consumer/join_type.rs
@@ -1,38 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-use datafusion::common::{plan_err, JoinType};
-use substrait::proto::join_rel;
-
-pub(super) fn from_substrait_jointype(
-    join_type: i32,
-) -> datafusion::common::Result<JoinType> {
-    if let Ok(substrait_join_type) = join_rel::JoinType::try_from(join_type) {
-        match substrait_join_type {
-            join_rel::JoinType::Inner => Ok(JoinType::Inner),
-            join_rel::JoinType::Left => Ok(JoinType::Left),
-            join_rel::JoinType::Right => Ok(JoinType::Right),
-            join_rel::JoinType::Outer => Ok(JoinType::Full),
-            join_rel::JoinType::LeftAnti => Ok(JoinType::LeftAnti),
-            join_rel::JoinType::LeftSemi => Ok(JoinType::LeftSemi),
-            join_rel::JoinType::LeftMark => Ok(JoinType::LeftMark),
-            _ => plan_err!("unsupported join type {substrait_join_type:?}"),
-        }
-    } else {
-        plan_err!("invalid join type variant {join_type:?}")
-    }
-}
