+++ datafusion/substrait/src/logical_plan/consumer/mod.rs
@@ -15,67 +15,16 @@
 // specific language governing permissions and limitations
 // under the License.
 
-mod agg_func;
-mod aggregate_rel;
-mod bound;
-mod cast;
-mod cross_rel;
-mod exchange_rel;
-mod extended_expr;
-mod fetch_rel;
-mod field_reference;
-mod filter_rel;
-mod func_args;
-mod grouping;
-mod if_then;
-mod join_rel;
-mod join_type;
-mod literal;
-mod named_struct;
 mod plan;
-mod project_rel;
-mod read_rel;
 mod rel;
 mod rex;
-mod rex_vec;
-mod scalar_function;
-mod set_rel;
-mod singular_or_list;
-mod sort;
-mod sort_rel;
-mod struct_type;
-mod subquery;
 mod substrait_consumer;
-mod r#type;
+mod types;
 mod utils;
-mod window_function;
 
-pub use agg_func::*;
-pub use aggregate_rel::*;
-pub use cast::*;
-pub use cross_rel::*;
-pub use exchange_rel::*;
-pub use extended_expr::*;
-pub use fetch_rel::*;
-pub use field_reference::*;
-pub use filter_rel::*;
-pub use func_args::*;
-pub use if_then::*;
-pub use join_rel::*;
-pub use literal::*;
-pub use named_struct::*;
 pub use plan::*;
-pub use project_rel::*;
-pub use r#type::*;
-pub use read_rel::*;
 pub use rel::*;
 pub use rex::*;
-pub use rex_vec::*;
-pub use scalar_function::*;
-pub use set_rel::*;
-pub use singular_or_list::*;
-pub use sort::*;
-pub use sort_rel::*;
-pub use subquery::*;
 pub use substrait_consumer::*;
-pub use window_function::*;
+pub use types::*;
+pub use utils::*;
