+++ datafusion/substrait/src/logical_plan/consumer/rel/exchange_rel.rs
@@ -15,8 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use super::from_substrait_field_reference;
-use super::SubstraitConsumer;
+use crate::logical_plan::consumer::from_substrait_field_reference;
+use crate::logical_plan::consumer::SubstraitConsumer;
 use datafusion::common::{not_impl_err, substrait_err};
 use datafusion::logical_expr::{LogicalPlan, Partitioning, Repartition};
 use std::sync::Arc;
