--- datafusion/substrait/src/logical_plan/consumer/sort.rs
@@ -1,72 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-use super::SubstraitConsumer;
-use datafusion::common::{not_impl_err, DFSchema};
-use datafusion::logical_expr::expr::Sort;
-use substrait::proto::sort_field::SortDirection;
-use substrait::proto::sort_field::SortKind::{ComparisonFunctionReference, Direction};
-use substrait::proto::SortField;
-
-/// Convert Substrait Sorts to DataFusion Exprs
-pub async fn from_substrait_sorts(
-    consumer: &impl SubstraitConsumer,
-    substrait_sorts: &Vec<SortField>,
-    input_schema: &DFSchema,
-) -> datafusion::common::Result<Vec<Sort>> {
-    let mut sorts: Vec<Sort> = vec![];
-    for s in substrait_sorts {
-        let expr = consumer
-            .consume_expression(s.expr.as_ref().unwrap(), input_schema)
-            .await?;
-        let asc_nullfirst = match &s.sort_kind {
-            Some(k) => match k {
-                Direction(d) => {
-                    let Ok(direction) = SortDirection::try_from(*d) else {
-                        return not_impl_err!(
-                            "Unsupported Substrait SortDirection value {d}"
-                        );
-                    };
-
-                    match direction {
-                        SortDirection::AscNullsFirst => Ok((true, true)),
-                        SortDirection::AscNullsLast => Ok((true, false)),
-                        SortDirection::DescNullsFirst => Ok((false, true)),
-                        SortDirection::DescNullsLast => Ok((false, false)),
-                        SortDirection::Clustered => not_impl_err!(
-                            "Sort with direction clustered is not yet supported"
-                        ),
-                        SortDirection::Unspecified => {
-                            not_impl_err!("Unspecified sort direction is invalid")
-                        }
-                    }
-                }
-                ComparisonFunctionReference(_) => not_impl_err!(
-                    "Sort using comparison function reference is not supported"
-                ),
-            },
-            None => not_impl_err!("Sort without sort kind is invalid"),
-        };
-        let (asc, nulls_first) = asc_nullfirst.unwrap();
-        sorts.push(Sort {
-            expr,
-            asc,
-            nulls_first,
-        });
-    }
-    Ok(sorts)
-}
