+++ datafusion/substrait/src/logical_plan/consumer/utils.rs
@@ -15,13 +15,19 @@
 // specific language governing permissions and limitations
 // under the License.
 
+use crate::logical_plan::consumer::SubstraitConsumer;
 use datafusion::arrow::datatypes::{DataType, Field, FieldRef, Schema};
 use datafusion::common::{
-    substrait_datafusion_err, substrait_err, DFSchema, DFSchemaRef, TableReference,
+    not_impl_err, substrait_datafusion_err, substrait_err, DFSchema, DFSchemaRef,
+    TableReference,
 };
+use datafusion::logical_expr::expr::Sort;
 use datafusion::logical_expr::{Cast, Expr, ExprSchemable, LogicalPlanBuilder};
 use std::collections::HashSet;
 use std::sync::Arc;
+use substrait::proto::sort_field::SortDirection;
+use substrait::proto::sort_field::SortKind::{ComparisonFunctionReference, Direction};
+use substrait::proto::SortField;
 
 // Substrait PrecisionTimestampTz indicates that the timestamp is relative to UTC, which
 // is the same as the expectation for any non-empty timezone in DF, so any non-empty timezone
@@ -322,6 +328,55 @@ impl NameTracker {
             (name, NameTrackerStatus::SeenBefore) => Ok(expr.alias(name)),
         }
     }
+}
+
+/// Convert Substrait Sorts to DataFusion Exprs
+pub async fn from_substrait_sorts(
+    consumer: &impl SubstraitConsumer,
+    substrait_sorts: &Vec<SortField>,
+    input_schema: &DFSchema,
+) -> datafusion::common::Result<Vec<Sort>> {
+    let mut sorts: Vec<Sort> = vec![];
+    for s in substrait_sorts {
+        let expr = consumer
+            .consume_expression(s.expr.as_ref().unwrap(), input_schema)
+            .await?;
+        let asc_nullfirst = match &s.sort_kind {
+            Some(k) => match k {
+                Direction(d) => {
+                    let Ok(direction) = SortDirection::try_from(*d) else {
+                        return not_impl_err!(
+                            "Unsupported Substrait SortDirection value {d}"
+                        );
+                    };
+
+                    match direction {
+                        SortDirection::AscNullsFirst => Ok((true, true)),
+                        SortDirection::AscNullsLast => Ok((true, false)),
+                        SortDirection::DescNullsFirst => Ok((false, true)),
+                        SortDirection::DescNullsLast => Ok((false, false)),
+                        SortDirection::Clustered => not_impl_err!(
+                            "Sort with direction clustered is not yet supported"
+                        ),
+                        SortDirection::Unspecified => {
+                            not_impl_err!("Unspecified sort direction is invalid")
+                        }
+                    }
+                }
+                ComparisonFunctionReference(_) => not_impl_err!(
+                    "Sort using comparison function reference is not supported"
+                ),
+            },
+            None => not_impl_err!("Sort without sort kind is invalid"),
+        };
+        let (asc, nulls_first) = asc_nullfirst.unwrap();
+        sorts.push(Sort {
+            expr,
+            asc,
+            nulls_first,
+        });
+    }
+    Ok(sorts)
 }
 
 #[cfg(test)]
